fix: 로그아웃 로직을 유틸함수->커스텀훅으로
